Ajustes Usuário controller e services. Refatoração do método: 1 - Update; 2 - Delete; 3 - listAll; 4 - findById. Nos métodos listados a cima incluímos Optional, validação e ajustes na pathVariable.

+++ src/main/java/com/sistema_campanha/controllers/UsuarioController.java
@@ -3,10 +3,14 @@ package com.sistema_campanha.controllers;
 import com.sistema_campanha.dto.UserDTO;
 import com.sistema_campanha.entities.Usuario;
 import com.sistema_campanha.services.UsuarioService;
+import jakarta.validation.Valid;
+import org.springframework.beans.BeanUtils;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
+import java.util.Optional;
 
 
 @RestController
@@ -20,28 +24,46 @@ public class UsuarioController {
 
 
     @GetMapping(value = "/{id}")
-    public UserDTO findById(@PathVariable Long id){
-        return service.findById(id);
+    public ResponseEntity<Object> getUsuarioId(@PathVariable(value = "id") Long id){
+        Optional<Usuario> usuario = service.findById(id);
+        if(usuario.isEmpty()){
+           return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Usuário não encontrado");
+        }
+        return ResponseEntity.status(HttpStatus.OK).body(usuario.get());
     }
 
     @GetMapping
-    public List<UserDTO> findAll(){
-        return service.getAll();
+    public ResponseEntity<List<Usuario>> findAll(){
+        return ResponseEntity.status(HttpStatus.OK).body(service.findAll());
     }
 
-    @PostMapping(path = "/usuario")
-    public Usuario saveUser(@RequestBody Usuario usuario){
-        return service.createdUser(usuario);
+    @PostMapping
+    public ResponseEntity<Usuario> saveUser(@RequestBody @Valid UserDTO userDTO){
+        var userModel = new Usuario();
+        BeanUtils.copyProperties(userDTO, userModel);
+        return ResponseEntity.status(HttpStatus.CREATED).body(service.createdUser(userModel));
     }
 
-    @PostMapping(value = "/delete/{id}")
-    public void deleteById(@PathVariable Long id){
+    @DeleteMapping(value = "/{id}")
+    public ResponseEntity<Object> deleteUser(@PathVariable(value = "id") Long id){
+        Optional<Usuario> optionalUsuario = service.findById(id);
+        if (optionalUsuario.isEmpty()){
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Usuário não encontrado!");
+        }
         service.deleteById(id);
+        return ResponseEntity.status(HttpStatus.OK).body("Usuário "+optionalUsuario.get().getNome()+" deletado com sucesso.");
     }
 
-    @PutMapping(value = "/update")
-    public Usuario update(@RequestBody Usuario usuario){
-        return service.update(usuario);
+    @PutMapping("/{id}")
+    public ResponseEntity<Object> updateUser(@PathVariable(value = "id") Long id,
+                                                @RequestBody @Valid UserDTO userDTO){
+        Optional<Usuario> usuario= service.findById(id);
+        if(usuario.isEmpty()){
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Product Não encontrado");
+        }
+        var userModel = usuario.get();
+        BeanUtils.copyProperties(userDTO, userModel);
+        return ResponseEntity.status(HttpStatus.OK).body(service.createdUser(userModel));
     }
